* adding custom element option to dialogs * allow coloring of textfields * displaying length of lines in select * postfix for textfields * default to PX_AND_SECONDS store in PhysicalProperties to more easily change cm_per_pixel in-App * adding alternative way of calibrating cm_per_pixel, resolves #24 * removing dead code and adding hint and error messages for python init (resolves #10)

+++ Application/src/tracker/misc/MemoryStats.cpp
@@ -197,7 +197,7 @@ uint64_t MemoryStats::get_memory_size(std::map<long_t, std::pair<void*, std::fun
     return sizeof(decltype(obj)::value_type) * obj.size();
 }
 
-MemoryStats::MemoryStats() : id(-1), bytes(0) {
+MemoryStats::MemoryStats() : id(uint32_t(-1)), bytes(0) {
     
 }
